move interface library to models folder
